Align UI closer to mockup provided by @kjell
See https://github.com/WordPress/gutenberg/issues/13363#issuecomment-463291100
